no err msg but start button not triggering func transition into Gameplay
